backend/requirements.txt
asgiref==3.10.0
Django==5.2.7
djangorestframework==3.16.1
psycopg2-binary==2.9.11
python-dotenv==1.2.1
sqlparse==0.5.3
typing_extensions==4.15.0
django-cors-headers==4.6.0
pytest==8.3.3
pytest-django==4.8.0
google-auth==2.25.2
djangorestframework-simplejwt==5.3.1
requests==2.31.0